requirements.txt
celery
redis
flower
pymongo
celery-redbeat
python-dotenv
web3
boto3